Move recordToSvgUpload function to recordHelpers to avoid duplicated code
