+++ src/entrypoints/PageScreen/PageScreen.tsx
@@ -13,35 +13,13 @@ import {
   createSvgRecord,
   updateSvgRecord,
   deleteSvgRecord,
+  recordToSvgUpload,
 } from '../../lib/recordHelpers'
 
 import * as styles from './PageScreen.module.css'
 
 type Props = {
   ctx: RenderPageCtx
-}
-
-// Helper to convert SvgRecord to SvgUpload format for compatibility with existing components
-function recordToSvgUpload(record: SvgRecord): SvgUpload {
-  const base = {
-    id: record.id,
-    filename: record.name || undefined,
-    raw: record.svg_content,
-  }
-
-  if (record.svg_type === 'image' && record.media_upload) {
-    return {
-      ...base,
-      type: 'image' as const,
-      imageId: record.media_upload.upload_id,
-      url: record.media_upload.url,
-    }
-  }
-
-  return {
-    ...base,
-    type: 'svg' as const,
-  }
 }
 
 export default function PageScreen({ ctx }: Props) {
